Refactor `Codec` implementations, migrate to `cryptic.codec.default`, and update test cases for consistency.
