+++ codec-upickle/src/test/scala/cryptic/codec/UpickleSpec.scala
@@ -3,9 +3,9 @@ package codec
 
 import org.scalatest.flatspec.AnyFlatSpec
 import org.scalatest.matchers.should.Matchers
-import upickle.default.*
+import upickle.default.{*, given}
 
-import scala.util.Success
+import scala.util.{Success, Try}
 
 case class Name(literal: String)
 
@@ -26,6 +26,7 @@ case class User(id: Long, alias: String, name: PersonName, email: EmailAddress)
 object User:
   implicit val rw: ReadWriter[User] = macroRW[User]
 class UpickleSpec extends AnyFlatSpec with Matchers:
+  import cryptic.codec.Upickle.{*, given}
   val user: User = User(
     id = 1,
     alias = "kalle",
@@ -34,20 +35,17 @@ class UpickleSpec extends AnyFlatSpec with Matchers:
   )
 
   "Upickle codec" should "encode string and then decode back to original string" in:
-    val codec: Codec[String] = Upickle[String]()
-    val plainText = codec.encode("kalle")
+    val plainText = "kalle".encoded
     plainText shouldNot equal(PlainText("kalle"))
-    val actual = codec.decode(plainText)
+    val actual: Try[String] = plainText.decoded
     actual shouldEqual Success("kalle")
   "Upickle codec" should "encode a case class and decode back to original case class" in:
-    val codec = Upickle[User]()
-    val plainText = codec.encode(user)
-    val actual = codec.decode(plainText)
+    val plainText = user.encoded
+    val actual: Try[User] = plainText.decoded
     actual shouldEqual Success(user)
   "Upickle codec" should "be usable by the encoder/decoder" in:
     import cryptic.crypto.Reverse.*
     import cryptic.crypto.Reverse.given
-    implicit val codec: Codec[User] = Upickle[User]()
     val encrypted = user.encrypted
     encrypted.bytes shouldBe write(user).getBytes.reverse // Reverse...
     val decrypted = encrypted.decrypted
