+++ core/src/test/scala/cryptic/crypto/ReverseSpec.scala
@@ -7,11 +7,12 @@ import org.scalatest.matchers.should.Matchers
 import scala.util.Success
 
 class ReverseSpec extends AnyFlatSpec with Matchers:
+  import cryptic.codec.default.{*, given}
   import Reverse.{*, given}
 
   private val text = "nisse"
   private val encrypted: Encrypted[String] = Encrypted(text)
-  
+
   "Reverse Encrypted" should "support encryption and decryption" in:
     encrypted.decrypted match
       case Success(decrypted) ⇒ decrypted shouldEqual text
@@ -20,6 +21,6 @@ class ReverseSpec extends AnyFlatSpec with Matchers:
   "Reverse Encrypted" should "hide plaintext" in:
     new String(encrypted.bytes)
       .contains(text) shouldBe false
-    
+
   "Reverse" should "be reversed" in:
     encrypted.bytes shouldEqual text.reverse.getBytes
