+++ crypto-test/src/test/scala/cryptic/test/bouncycastle/ECSpecBase.scala
@@ -6,10 +6,8 @@ import cryptic.{Decrypt, Encrypt}
 import cryptic.crypto.EC
 import cryptic.codec.{Chill, Fst, Upickle}
 import cryptic.test.CryptoSpecBase
-import upickle.default.*
-
+import upickle.default.{ReadWriter, Writer, Reader, given}
 import java.security.{KeyPair, PrivateKey, PublicKey}
-import scala.compiletime.deferred
 
 trait ECSpecBase extends CryptoSpecBase:
   private val keyPair: KeyPair = EC.keygen(256)
@@ -19,14 +17,13 @@ trait ECSpecBase extends CryptoSpecBase:
   override given decrypt: Decrypt = EC.decrypt
 
 class ECChillSpec extends ECSpecBase:
-  override given codec[V]: Codec[V] = Chill.codec
+  override given stringCodec: Codec[String] = Chill.codec
   override def toString: String = "ECChill"
 
 class ECFstSpec extends ECSpecBase:
-  override given codec[V]: Codec[V] = Fst.codec
+  override given stringCodec: Codec[String] = Fst.codec
   override def toString: String = "ECFst"
 
-class ECUpickleSpec[V: ReadWriter] extends ECSpecBase:
-  override given codec[W]: Codec[W] =
-    Upickle[W]()(using summon[ReadWriter[V]].asInstanceOf[ReadWriter[W]])
+class ECUpickleSpec extends ECSpecBase:
+  override given stringCodec: Codec[String] = Upickle.codec[String]
   override def toString: String = "ECUpickle"
